requirements.txt
# Django libraries
django
djangorestframework
django-debug-toolbar
django-extensions
djangorestframework-simplejwt[crypto]
django-cors-headers
Werkzeug

# Tests libraries
faker
factory_boy

# Other librairies
markdown
